Move FilterButton component to separate file
